Use surrounding points for joint calibration. Remove memory allocation from interpolation function.
