+++ sr_robot_lib/src/sr_motor_robot_lib.cpp
@@ -973,22 +973,18 @@ namespace shadow_robot
           int raw_pos_2nd_joint = status_data->sensors[this->joints_vector[this->find_joint_by_name("THJ2")].joint_to_sensor.joint_to_sensor_vector[0].sensor_id];
           this->xyi[0] = static_cast<double> (raw_pos);
           this->xyi[1] = static_cast<double> (raw_pos_2nd_joint);
-          this->zi = pwl_interp_2d_scattered_value (this->node_num, this->node_xy, this->zd_thj1, this->element_num,
-                                                    this->triangle, this->element_neighbor, this->ni, this->xyi);
+          pwl_interp_2d_scattered_value (this->node_num, this->node_xy, this->zd_thj1, this->element_num,
+                                         this->triangle, this->element_neighbor, this->ni, this->xyi, this->zi);
           tmp_cal_value = this->zi[0];
-          // ROS_INFO("THJ1: %f THJ2: %f Interpolated THJ1: %f", this->xyi[0], this->xyi[1], tmp_cal_value);
-          delete [] this->zi;
         }
         else if (joint_tmp->joint_name.find("THJ2")!= std::string::npos)
         {
           int raw_pos_2nd_joint = status_data->sensors[this->joints_vector[this->find_joint_by_name("THJ1")].joint_to_sensor.joint_to_sensor_vector[0].sensor_id];
           this->xyi[0] = static_cast<double> (raw_pos_2nd_joint);
           this->xyi[1] = static_cast<double> (raw_pos);
-          this->zi = pwl_interp_2d_scattered_value (this->node_num, this->node_xy, this->zd_thj2, this->element_num,
-                                                    this->triangle, this->element_neighbor, this->ni, this->xyi);
+          pwl_interp_2d_scattered_value (this->node_num, this->node_xy, this->zd_thj2, this->element_num,
+                                         this->triangle, this->element_neighbor, this->ni, this->xyi, this->zi);
           tmp_cal_value = this->zi[0];
-          // ROS_INFO("THJ1: %f THJ2: %f Interpolated THJ2: %f", this->xyi[0], this->xyi[1], tmp_cal_value);
-          delete [] this->zi;
         }
         else
         {
